requirements.txt
Flask
dnspython
requests
flask
